Rename `utils.py` → `_utils.py`
