+++ qtpy/_utils.py
